refactor: 이름을 통일감있게 변경.
